+++ apps/app/components/issues/select/priority.tsx
@@ -2,9 +2,10 @@ import React from "react";
 
 // headless ui
 import { Listbox, Transition } from "@headlessui/react";
+// icons
+import { getPriorityIcon } from "components/icons/priority-icon";
 // constants
-import { getPriorityIcon } from "constants/global";
-import { PRIORITIES } from "constants/";
+import { PRIORITIES } from "constants/project";
 
 type Props = {
   value: string | null;
