+++ apps/app/constants/module.ts
@@ -1,0 +1,8 @@
+export const MODULE_STATUS = [
+  { label: "Backlog", value: "backlog", color: "#5e6ad2" },
+  { label: "Planned", value: "planned", color: "#26b5ce" },
+  { label: "In Progress", value: "in-progress", color: "#f2c94c" },
+  { label: "Paused", value: "paused", color: "#ff6900" },
+  { label: "Completed", value: "completed", color: "#4cb782" },
+  { label: "Cancelled", value: "cancelled", color: "#cc1d10" },
+];
