+++ apps/app/constants/project.ts
@@ -1,0 +1,11 @@
+export const NETWORK_CHOICES = { "0": "Secret", "2": "Public" };
+
+export const GROUP_CHOICES = {
+  backlog: "Backlog",
+  unstarted: "Unstarted",
+  started: "Started",
+  completed: "Completed",
+  cancelled: "Cancelled",
+};
+
+export const PRIORITIES = ["urgent", "high", "medium", "low", null];
