+++ apps/app/hooks/use-my-issues-filter.tsx
@@ -13,7 +13,7 @@ import { Properties, NestedKeyOf, IIssue } from "types";
 // fetch-keys
 import { STATE_LIST } from "constants/fetch-keys";
 // constants
-import { PRIORITIES } from "constants/";
+import { PRIORITIES } from "constants/project";
 
 const initialValues: Properties = {
   key: true,
