+++ apps/app/pages/[workspaceSlug]/settings/index.tsx
@@ -35,7 +35,7 @@ import type { GetServerSideProps, NextPage } from "next";
 // fetch-keys
 import { WORKSPACE_DETAILS, USER_WORKSPACES } from "constants/fetch-keys";
 // constants
-import { companySize } from "constants/";
+import { COMPANY_SIZE } from "constants/workspace";
 
 const defaultValues: Partial<IWorkspace> = {
   name: "",
@@ -278,7 +278,7 @@ const WorkspaceSettings: NextPage<TWorkspaceSettingsProps> = (props) => {
                     label={value ? value.toString() : "Select company size"}
                     input
                   >
-                    {companySize?.map((item) => (
+                    {COMPANY_SIZE?.map((item) => (
                       <CustomSelect.Option key={item.value} value={item.value}>
                         {item.label}
                       </CustomSelect.Option>
